fix: 필터 연관 키워드 티커, 테그 API DDD + 핵사고날로 리팩토링
